Agregando funciones de hash + fuente.
Tambien cambie un par de cosas porque soy un gede (ej: v1->vector1) para que queden mas lindos, y cambie los vectores a size_t (nunca van a tener numeros negativos, y asi entran el doble de numeros).

+++ minsketch.h
@@ -20,7 +20,7 @@ void min_guardar(minsketch_t* min, char* clave);
  * Si devuelve 0, quiere decir que nunca fue guardada.
  * Pre: El TDA fue creado.
  */
-int min_obtener(minsketch_t* min, const char* clave);
+size_t min_obtener(minsketch_t* min, const char* clave);
 
 /* Destruye el minsketch pasado por parametro.
  * Pre: El TDA fue creado.
